challenge/eth/requirements.txt
web3>=6.0.0
py-solc-x>=2.0.0
